+++ tools/libexec/perl/TTP/IFindable.pm
@@ -193,6 +193,12 @@ after _newBase => sub {
 	$self->{_ifindable} //= {};
 };
 
+### Global functions
+### Note for the developer: while a global function doesn't take any argument, it can be called both
+### as a class method 'TTP::Package->method()' or as a global function 'TTP::Package::method()',
+### the former being preferred (hence the writing inside of the 'Class methods' block which brings
+### the class as first argument).
+
 1;
 
 __END__